refactor: move files around pt2
- remame classes that have moved
