+++ src/components/SendPaymentPopup.tsx
@@ -7,7 +7,7 @@ import { useWalletStore } from "../store/wallet.store";
 import { useMessagingStore } from "../store/messaging.store";
 import { encrypt_message } from "cipher";
 import { Address } from "kaspa-wasm";
-import { toast } from "../utils/toast";
+import { toast } from "../utils/toast-helper";
 import { KasiaTransaction } from "../types/all";
 import { PROTOCOL } from "../config/protocol";
